@@ -6,17 +6,18 @@
   "author": "James Gillmore <james@faceyspacey.com>",
   "license": "MIT",
   "scripts": {
-    "start": "npm run clean && cross-env NODE_ENV=development node server/index.js",
-    "start:prod": "npm run clean && cross-env NODE_ENV=production node server/index.js",
+    "start": "npm run clean && cross-env NODE_ENV=development babel-node --extensions '.ts,.tsx' server/index.ts",
+    "start:prod": "npm run clean && cross-env NODE_ENV=production babel-node --extensions '.ts,.tsx' server/index.ts",
     "format": "prettier --write '{src,server,webpack}/**/*.js'",
     "clean": "rimraf buildClient buildServer"
   },
   "dependencies": {
-    "@babel/preset-stage-3": "^7.0.0-beta.39",
+    "@babel/node": "^7.0.0-beta.40",
+    "@babel/preset-stage-3": "^7.0.0-beta.40",
     "@types/compression": "0.0.35",
     "@types/express": "^4.11.1",
-    "@types/webpack": "^3.8.5",
-    "@types/ws": "^4.0.0",
+    "@types/webpack": "^3.8.8",
+    "@types/ws": "^4.0.1",
     "colors": "^1.1.2",
     "compression": "^1.7.1",
     "enzyme": "^3.3.0",
@@ -29,8 +30,8 @@
     "react-dnd": "^2.5.4",
     "react-dnd-html5-backend": "^2.5.4",
     "react-dom": "^16.2.0",
-    "react-markdown": "^3.1.5",
-    "react-redux": "^5.0.6",
+    "react-markdown": "^3.2.0",
+    "react-redux": "^5.0.7",
     "react-universal-component": "^2.8.2",
     "recompose": "^0.26.0",
     "reconnectingwebsocket": "^1.0.0",
@@ -41,35 +42,35 @@
     "ws": "^4.0.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.0.0-beta.39",
-    "@babel/core": "^7.0.0-beta.39",
-    "@babel/plugin-proposal-class-properties": "^7.0.0-beta.39",
-    "@babel/plugin-syntax-dynamic-import": "^7.0.0-beta.39",
-    "@babel/preset-env": "^7.0.0-beta.39",
-    "@babel/preset-react": "^7.0.0-beta.39",
-    "@babel/preset-typescript": "^7.0.0-beta.39",
-    "@types/enzyme": "^3.1.8",
-    "@types/lodash": "^4.14.102",
-    "@types/radium": "^0.18.23",
-    "@types/react": "^16.0.36",
-    "@types/react-dnd": "^2.0.34",
-    "@types/react-dnd-html5-backend": "^2.1.8",
-    "@types/react-dom": "^16.0.3",
-    "@types/react-hot-loader": "^3.0.5",
-    "@types/react-redux": "^5.0.14",
-    "@types/recompose": "^0.24.4",
+    "@babel/cli": "^7.0.0-beta.40",
+    "@babel/core": "^7.0.0-beta.40",
+    "@babel/plugin-proposal-class-properties": "^7.0.0-beta.40",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0-beta.40",
+    "@babel/preset-env": "^7.0.0-beta.40",
+    "@babel/preset-react": "^7.0.0-beta.40",
+    "@babel/preset-typescript": "^7.0.0-beta.40",
+    "@types/enzyme": "^3.1.9",
+    "@types/lodash": "^4.14.104",
+    "@types/radium": "^0.18.24",
+    "@types/react": "^16.0.38",
+    "@types/react-dnd": "^2.0.36",
+    "@types/react-dnd-html5-backend": "^2.1.9",
+    "@types/react-dom": "^16.0.4",
+    "@types/react-hot-loader": "^3.0.6",
+    "@types/react-redux": "^5.0.15",
+    "@types/recompose": "^0.24.5",
     "babel-loader": "v8.0.0-beta.0",
     "babel-plugin-universal-import": "^1.4.0",
     "babel-watch": "^2.0.7",
     "cross-env": "^5.1.3",
     "css-loader": "^0.28.9",
-    "eslint": "^4.17.0",
+    "eslint": "^4.18.0",
     "eslint-plugin-import": "^2.8.0",
     "extract-css-chunks-webpack-plugin": "^2.0.18",
     "prettier": "^1.10.2",
     "react-hot-loader": "^3.1.3",
     "rimraf": "^2.6.2",
-    "webpack": "^3.10.0",
+    "webpack": "^3.11.0",
     "webpack-hot-middleware": "^2.21.0",
     "webpack-hot-server-middleware": "^0.4.0"
   }
